+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Trulioo SDK for Java Changelog
 
+## Version 3.0.4
+Added new document types
+- CompletePlus
+- ArticleOfAuthority
+- AgentAddressChange
+
+## Version 3.0.3
+Added new document types
+- RegisterReport
+- CreditCheck
+- CreditReport
+- GISAExtract
+- VRExtract
+- RegisterCheck
+- TradeRegisterReport
+- BeneficialOwnersCheck
+- AnnualAccounts
+- FiledChanges
+- FiledDocuments
+
 # Version 3.0.2
 
 Simplify names of getAllDatasources methods.